pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "bforbuntyai"
version = "0.1.0"
description = "Run GenAI experiments in a few lines instead of hundreds"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["genai", "deep-learning", "gan", "vae", "diffusion", "transformers"]
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "tqdm>=4.65",
    "Pillow>=10.0",
    "requests>=2.31",
    "huggingface_hub>=0.20",
]

[project.optional-dependencies]
torch = [
    "torch>=2.0",
    "torchvision>=0.15",
]
tensorflow = [
    "tensorflow>=2.13",
]
transformers = [
    "transformers>=4.35",
    "datasets>=2.14",
    "accelerate>=0.24",
]
diffusers = [
    "diffusers>=0.23",
    "transformers>=4.35",
    "accelerate>=0.24",
]
ethics = [
    "detoxify>=0.5",
    "transformers>=4.35",
]
all = [
    "torch>=2.0",
    "torchvision>=0.15",
    "tensorflow>=2.13",
    "transformers>=4.35",
    "datasets>=2.14",
    "accelerate>=0.24",
    "diffusers>=0.23",
    "detoxify>=0.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bforbuntyai*"]

[tool.setuptools.package-data]
bforbuntyai = ["py.typed"]
